pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["nhirzel"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
